requirements.txt
requests
python-dotenv
groq
langchain
spacy
langchain_community
flask
flask_cors
openai

